Rename merger.h -> merging_iterator.h
