+++ src.mk
@@ -74,7 +74,7 @@ LIB_SOURCES =                                                   \
   table/full_filter_block.cc                                    \
   table/get_context.cc                                          \
   table/iterator.cc                                             \
-  table/merger.cc                                               \
+  table/merging_iterator.cc                                     \
   table/meta_blocks.cc                                          \
   table/sst_file_writer.cc                                      \
   table/plain_table_builder.cc                                  \
