+++ table/merging_iterator.cc
@@ -7,7 +7,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
-#include "table/merger.h"
+#include "table/merging_iterator.h"
 #include <string>
 #include <vector>
 #include "db/pinned_iterators_manager.h"
